break main app into components
