+++ src/MarkdownEditor.js
@@ -14,6 +14,11 @@ var MarkdownSelectionStore = require('./stores/MarkdownSelectionStore');
 var MarkdownEditorTabsInteractionStore = require('./stores/MarkdownEditorTabsInteractionStore');
 var MarkdownTokenFactory = require('./utils/MarkdownTokenFactory');
 var MarkdownUtils = require('./utils/MarkdownUtils');
+var MarkdownEditorMenu = require('./compnents/MarkdownEditorMenu');
+var MarkdownEditorTabs = require('./compnents/MarkdownEditorTabs');
+var MarkdownEditorContent = require('./compnents/MarkdownEditorContent');
+var MarkdownEditorPreview = require('./compnents/MarkdownEditorPreview');
+var MarkdownEditor = require('./compnents/MarkdownEditor');
 
 var NullMarkdownToken = MarkdownTokenFactory.NullMarkdownToken;
 var RegularMarkdownToken = MarkdownTokenFactory.RegularMarkdownToken;
@@ -22,249 +27,6 @@ var SubHeaderMarkdownToken = MarkdownTokenFactory.SubHeaderMarkdownToken;
 var UrlMarkdownToken = MarkdownTokenFactory.UrlMarkdownToken;
 var ListMarkdownToken = MarkdownTokenFactory.ListMarkdownToken;
 var ImageMarkdownToken = MarkdownTokenFactory.ImageMarkdownToken;
-
-var MarkdownEditorMenu = React.createClass({
-  mixins: [Reflux.ListenerMixin, ButtonManagerMixin],
-
-  propTypes: {
-    iconsSet: React.PropTypes.string.isRequired
-  },
-
-  getInitialState: function() {
-    return {
-      enabled: false
-    };
-  },
-
-  componentWillMount: function() {
-    this.listenTo(MarkdownSelectionStore, this.handleMarkdownSelectionStore);
-    this.setIconsProvider(this.props.iconsSet);
-  },
-
-  render: function() {
-    var styleMarkdownMenu = {
-      'margin': '5px 0',
-      'flex': '1',
-      'display': 'flex',
-      'position': 'absolute',
-      'right': '20px',
-      'top': '10px'
-    };
-
-    var _disabled = (!this.state.enabled) ? 'disabled' : '';
-    var boldButton = this.getBoldButton(_disabled, this.handleBoldButtonClick);
-    var italicButton = this.getItalicButton(_disabled, this.handleItalicButtonClick);
-    var makeListButton = this.getMakeListButton(_disabled, this.handleListButtonClick);
-    var imageButton = this.getImageButton(_disabled, this.handleImageButtonClick);
-    var linkButton = this.getLinkButton(_disabled, this.handleLinkButtonClick);
-    var headerButton = this.getButtonWithoutIcon(_disabled, this.handleHeaderButtonClick, 'md-editor-menu-header', 'Header');
-    var subHeaderButton = this.getButtonWithoutIcon(_disabled, this.handleSubHeaderButtonClick, 'md-editor-menu-subheader', 'Subheader');
-
-    return (
-      <div style={styleMarkdownMenu} className='md-editor-menu'>
-        {boldButton}
-        {italicButton}
-        {headerButton}
-        {subHeaderButton}
-        {makeListButton}
-        {imageButton}
-        {linkButton}
-      </div>
-    );
-  },
-
-  handleMarkdownSelectionStore: function(data) {
-    if (data.type === 'clear') {
-      this.setState({enabled: false});
-    } else if (data.type === 'set') {
-      this.setState({enabled: true});
-    }
-  },
-
-  handleBoldButtonClick: function() {
-    MarkdownEditorActions.makeBold();
-  },
-
-  handleImageButtonClick: function() {
-    MarkdownEditorActions.makeImage();
-  },
-
-  handleItalicButtonClick: function() {
-    MarkdownEditorActions.makeItalic();
-  },
-
-  handleUnderlineButtonClick: function() {
-    MarkdownEditorActions.makeUnderline();
-  },
-
-  handleHeaderButtonClick: function() {
-    MarkdownEditorActions.makeHeader();
-  },
-
-  handleSubHeaderButtonClick: function() {
-    MarkdownEditorActions.makeSubHeader();
-  },
-
-  handleLinkButtonClick: function() {
-    MarkdownEditorActions.makeLink();
-  },
-
-  handleListButtonClick: function() {
-    MarkdownEditorActions.makeList();
-  }
-});
-
-var MarkdownEditorTabs = React.createClass({
-  mixins: [Reflux.ListenerMixin],
-
-  getInitialState: function() {
-    return {
-      activeTab: 0
-    };
-  },
-
-  componentWillMount: function() {
-    this.listenTo(MarkdownEditorTabsInteractionStore, this.handleMDEditorTabsInteractionStoreUpdated);
-  },
-
-  handleMDEditorTabsInteractionStoreUpdated: function(storeState) {
-    if (storeState.activeTab != null) {
-      this.setState({activeTab: storeState.activeTab});
-    }
-  },
-
-  render: function() {
-    var styleMarkdownEditorTabs = {
-      'border': 'none',
-      'display': 'flex',
-      'justifyContent': 'flex-start'
-    };
-
-    var styleTab = {
-      'padding': '0px 20px',
-      'cursor': 'pointer',
-      'display': 'flex',
-      'justifyContent': 'center',
-      'alignItems': 'center',
-      'height': '50px'
-    };
-
-    var styleActiveTab = {
-      'padding': '0px 20px',
-      'cursor': 'pointer',
-      'display': 'flex',
-      'justifyContent': 'center',
-      'alignItems': 'center',
-      'height': '50px',
-      'borderLeft': '1px solid #ddd',
-      'borderRight': '1px solid #ddd',
-      'borderTop': '1px solid #ddd',
-      'backgroundColor': '#fff',
-      'borderRadius': '3px'
-    };
-
-    var editorTabStyle;
-    var previewTabStyle;
-    if (this.state.activeTab === 0) {
-      editorTabStyle = styleActiveTab;
-      previewTabStyle = styleTab;
-    } else if (this.state.activeTab === 1) {
-      previewTabStyle = styleActiveTab;
-      editorTabStyle = styleTab;
-    }
-
-    return (
-      <div style={styleMarkdownEditorTabs} className='md-editor-tabs'>
-        <div style={editorTabStyle}
-          className="md-editor-tabs-item"
-          onClick={this.handleClick.bind(this, 'clickEditorTab')}>
-          <span>Editor</span>
-        </div>
-        <div style={previewTabStyle}
-          className="md-editor-tabs-item"
-          onClick={this.handleClick.bind(this, 'clickPreviewTab')}>
-          <span>Preview</span>
-        </div>
-      </div>
-    );
-  },
-
-  handleClick: function(actionName) {
-    MarkdownEditorActions[actionName]();
-  }
-});
-
-var MarkdownEditorContent = React.createClass({
-  propTypes: {
-    content: React.PropTypes.string.isRequired,
-    onChangeHandler: React.PropTypes.func.isRequired
-  },
-
-  mixins: [TextAreaSelectionMixin],
-
-  render: function() {
-    var styleMarkdownTextArea = {
-      'height': '90%',
-      'width': '100%',
-      'padding': '30px 10px',
-      'border': 'none'
-    };
-
-    return (
-      <textarea
-        ref='editor'
-        className='md-editor-textarea'
-        style={styleMarkdownTextArea}
-        onChange={this.onChange}
-        onClick={this.clearSelection}
-        onKeyUp={this.clearSelection}>
-      </textarea>
-    );
-  },
-
-  onChange: function() {
-    var content = this.refs.editor.value;
-    var markdownContent = MarkdownUtils.toMarkdown(content);
-    PublicMarkdownEditorActions.updateText(markdownContent);
-
-    this.props.onChangeHandler(content.replace(/[\n\r]/g, '\n'));
-  },
-
-  componentDidMount: function() {
-    this.refs.editor.value = this.props.content;
-  },
-
-  componentDidUpdate: function() {
-    this.refs.editor.value = this.props.content;
-  }
-});
-
-var MarkdownEditorPreview = React.createClass({
-  propTypes: {
-    content: React.PropTypes.string.isRequired
-  },
-
-  render: function() {
-    // Breaklines in markdown are actually when a line is ended with two spaces + carriage-return
-    var htmlContent = this.props.content.replace(/[\n]/g, '  \n');
-    htmlContent = Markdown.toHTML(htmlContent);
-
-    var styleMarkdownPreviewArea = {
-      'height': '90%',
-      'width': '100%',
-      'padding': '30px 10px',
-      'backgroundColor': '#fff',
-      'border': 'none'
-    };
-
-    return (
-      <div
-        style={styleMarkdownPreviewArea}
-        dangerouslySetInnerHTML={{__html: htmlContent}}
-      />
-    );
-  }
-});
 
 var MarkdownEditor = React.createClass({
   mixins: [Reflux.ListenerMixin],
@@ -302,6 +64,10 @@ var MarkdownEditor = React.createClass({
       'position': 'relative'
     };
 
+    if(this.props.styles.styleMarkdownEditorHeader) {
+      styleMarkdownEditorHeader = this.props.styles.styleMarkdownEditorHeader;
+    }
+
     var styleMarkdownEditorContainer = {
       'display': 'flex',
       'flexDirection': 'column',
@@ -310,6 +76,10 @@ var MarkdownEditor = React.createClass({
       'border': '1px solid #ddd',
       'backgroundColor': '#f7f7f7'
     };
+
+    if(this.props.styles.styleMarkdownEditorContainer) {
+      styleMarkdownEditorContainer = this.props.styles.styleMarkdownEditorContainer;
+    }
 
     return (
       <div
